urban_clap/requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.
pillow==10.3.0
